backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# Auth
pyjwt==2.10.1
bcrypt==4.2.1
cryptography==44.0.0

# LLM
litellm==1.55.8
httpx>=0.27.0

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# WebSocket
websockets==14.1

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2
pyyaml==6.0.2

# Utils
python-jose[cryptography]==3.3.0
